data/pyproject.toml
[project]
name = "cp-bg-bench"
version = "0.1.0"
description = "Cell Painting biology-grounded benchmark: unified segmentation + crop dataset pipeline."
requires-python = ">=3.11"
license = { text = "MIT" }
authors = ["Anonymous <anon@example.invalid>"]
dependencies = [
    "pydantic>=2",
    "pyyaml",
    "pandas",
    "pyarrow",
    "zarr>=3",
    "numcodecs",
    "tqdm",
    "lamin_utils",
]

[project.optional-dependencies]
io = ["boto3", "aiobotocore", "tifffile", "fastparquet"]
gpu = ["torch", "cellpose>=4", "opencv-python-headless", "cupy-cuda12x"]
hf = ["datasets", "dask[distributed]"]
dev = ["pytest", "ruff", "mypy"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cp_bg_bench"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
